backend/requirements.txt
channels==3.0.4
django-cors-headers==4.0.0
djangorestframework==3.14.0
asgiref==3.7.2
Django==4.2.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
easy-thumbnails==2.8.5